my355: align stock OS hook with Spruce

+++ skeleton/BOOT/miyoo355/app/my355.sh
@@ -5,12 +5,11 @@ cd $(dirname "$0")
 SDCARD_PATH=/mnt/SDCARD
 MIYOO_DIR=miyoo355
 
-if [ ! -f /usr/miyoo/bin/runmiyoo-original.sh ]; then
-	if [ -d $SDCARD_PATH/.tmp_update/ ]; then
-		rm -rf $SDCARD_PATH/.tmp_update/
-	fi
-	my355/init.sh
+if [ ! -f /usr/miyoo/bin/runmiyoo-original.sh ] && [ -d $SDCARD_PATH/.tmp_update/ ]; then
+	rm -rf $SDCARD_PATH/.tmp_update/
 fi
+# installs the hook, or upgrades it if ours is newer
+my355/init.sh
 
 export PATH=/usr/miyoo/bin:/usr/miyoo/sbin:/usr/bin:/usr/sbin:/bin:/sbin
 export LD_LIBRARY_PATH=/usr/miyoo/lib:/usr/lib:/lib
